feat: bunch of additional documentation
